Signatures on gun-related petitions
Individual signatures on gun-related petitions up to the end of May, 2016.  In the middle of June, 2016 the petition server API stopped making public the zip code and initials of signators.
Data include the zipcode, initials, state (as entered by signator), state (as determined from zipcode), petition ID, and whether the petition was pro- or anti-gun control.